% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map visualising all accident events in a given State for a paricular year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a State Code Number as Integer}

\item{year}{a given Year as Integer}
}
\value{
This function returns a character vector  the format "accident_YEAR.csv.bz2"
}
\description{
This function takes a State number (\code{state.num}) indicator and a year (\code{year}).
}
\details{
If the State number does not exhist in the data for the given year, the method exits (STOP)
with a message "invalid STATE number: ???".

If no data for State Num / year combination a message "no accidents to plot" is returned.
}
\examples{
\dontrun{
fars_map_state(12, 2017)
}

}
