% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Verify if the file exists using the \code{filename} function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file; provided as character string.}
}
\value{
dataframe representation of the file contents defined filename or
'does not exist' message.
}
\description{
Verify if the file exists using the \code{filename} function
}
\examples{
\dontrun{
fars_read("filename.csv")
fars_read("path/filename.csv")
}

}
