% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple .csv data file for a series a provided years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers representing a series of years}
}
\value{
This function returns a dataframe with Year and Months columns for each year.
}
\description{
This function take as input a vector of integers (\code{years}); returns a list collection of
DataFrames of Years and respesctive months.
}
\details{
For each year with no data file, the function generates a warning message.
}
\examples{
\dontrun{
fars_read_years(2017)
fars_read_years(2015:2016)
}

}
