% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise data on month and given year from data format defined in "accident_YEAR.csv.bz2" files.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years given as integers.}
}
\value{
This function returns a dataframe with Year and Months columns for each year.
}
\description{
This function takes a collection of years (\code{years}) as input and returns a Tibble(Data Frame)
containing the defined data summary as a Wide Format.
}
\details{
For each year with no data file, the function generates a warning message.
}
\examples{
\dontrun{
fars_summarize_years(2017)
fars_summarize_years(2015:2016)
}

}
